FRONT-42 Fetch data on first load
